Support for efficient reading directly from file

+++ src/main/java/me/doubledutch/lazyjson/LazyException.java
@@ -4,7 +4,7 @@ package me.doubledutch.lazyjson;
  * Exception used to indicate a parse or access error for LazyObject and LazyArray
  */
 public final class LazyException extends RuntimeException{
-	private int position=-1;
+	private long position=-1;
 	private String message;
 	
 	public LazyException(String str){
@@ -12,7 +12,7 @@ public final class LazyException extends RuntimeException{
 		this.message=str;
 	}
 
-	public LazyException(String str,int position){
+	public LazyException(String str,long position){
 		super(str);
 		this.position=position;
 		this.message=str;
